agent/requirements-e2e.txt
# Playwright UI E2E (CI job e2e-ui + local: pip install -r requirements-e2e.txt && playwright install chromium)
pytest-playwright>=0.5.0
playwright>=1.49.0
